split agenda.py
