+++ app/views.py
@@ -4,18 +4,9 @@ import requests
 from django.shortcuts import render
 from django.http import HttpResponse
 from django.http import HttpResponseRedirect
-from django.http import QueryDict
-from django.shortcuts import render_to_response
-from django.template import RequestContext
 
 from .models import Greeting
 from .models import FikanoteDB, AgendaDB, Shownote
-
-from agendaform import AgendaForm
-
-import urllib2
-from BeautifulSoup import BeautifulSoup
-import datetime
 
 def index(request):
     episodes = FikanoteDB.objects.order_by('-date')
@@ -38,46 +29,6 @@ def episode(request, number):
                    'person' : person,
                    'shownotes' : shownotes,
                    } )
-
-def agenda(request):
-    import json
-    from django.http import HttpResponse,Http404
-
-    if request.method == 'GET': 
-        agendas = AgendaDB.objects().order_by('-date')
-        form = AgendaForm() 
-        return render(request, 'agenda.html', 
-                      {'agendas': agendas, 
-                       'form': form
-                       } )
-
-    elif request.method == 'POST': 
-        form = AgendaForm(request.POST) 
-        if form.is_valid(): 
-            url = form.cleaned_data['url']
-            req = urllib2.Request(url, None, headers = { 'User-Agent' : 'Mozilla/5.0' })
-            soup = BeautifulSoup(urllib2.urlopen(req))
-            title = soup.title.string
-            AgendaDB(url=url, title=title, date=datetime.datetime.utcnow()).save()
-
-            response = json.dumps({'status':'success', 'url':url, 'title':title})  # convert to JSON
-        else:
-            response = json.dumps({'status':'fail'})  # convert to JSON
-
-        return HttpResponse(response, content_type="application/json")
-
-    elif request.method == 'DELETE': 
-        delete = QueryDict(request.body)
-        id = delete.get('id')
-        if id :
-            AgendaDB.objects.filter(id__exact=id).delete()
-            response = json.dumps({'status':'success'})  # convert to JSON
-        else:
-            response = json.dumps({'status':'fail'})  # convert to JSON
-        return HttpResponse(response, content_type="application/json")
-
-    else:
-        raise Http404
 
 
 def add(request):
