+++ fikanote/urls.py
@@ -6,6 +6,7 @@ admin.autodiscover()
 import app.views
 import app.feed
 import app.shownote
+import app.agenda
  
 urlpatterns = patterns('',
     # Examples:
@@ -14,7 +15,7 @@ urlpatterns = patterns('',
 
     url(r'^$', app.views.index, name='index'),
     url(r'^(?P<number>\d+)', app.views.episode, name='episode'),
-    url(r'^agenda', app.views.agenda, name='agenda'),
+    url(r'^agenda', app.agenda.agenda, name='agenda'),
     url(r'^add', app.views.add, name='add'),
     url(r'^shownote', app.shownote.shownote, name='shownote'),
     url(r'^feed/', app.feed.feed, name='feed'),
